remove duplicate dir, fix module imports and typos
