+++ src/Augraphy/augmentations/brightnesstexturize.py
@@ -18,6 +18,8 @@ import cv2
 import numpy as np
 import random
 
+from Augraphy.Augmentations import *
+
 
 
 ################################################################################
