+++ src/Augraphy/augmentations/noisetexturize.py
@@ -17,6 +17,8 @@ import cv2
 import numpy as np
 import random
 
+from Augraphy.Augmentations import *
+
 
 
 ################################################################################
