+++ src/Augraphy/augmentations/subtlenoise.py
@@ -1,10 +1,11 @@
 ################################################################################
-# File: jpeg.py
+# File: subtlenoise.py
 #
-# Class: JpegAugmentation
+# Class: SubtleNoiseAugmentation
 #
-# Description: This file contains a class defining an Augmentation which uses
-#              JPEG encoding to create JPEG compression artifacts in the image.
+# Description: This file contains a class defining an Augmentation which
+#              emulates the imperfections in scanning solid colors due to subtle
+#              lighting differences.adjusts
 ################################################################################
 
 
@@ -13,8 +14,10 @@
 # Imports
 ################################################################################
 
-import cv2
-import numpy.random
+import numpy as np
+import random
+
+from Augraphy.Augmentations import *
 
 
 
@@ -22,20 +25,19 @@ import numpy.random
 # Class Definition
 ################################################################################
 
-class JpegAugmentation(Augmentation):
-  def __init__(self, quality_range=(25, 95), probability=0.5):
+class SubtleNoiseAugmentation(Augmentation):
+  def __init__(self, range=10, probability=0.5):
     super().__init__(probability=probability)
-    self.quality_range = quality_range
+    self.range = range
+    self.add_subtle_noise = np.vectorize(lambda x: max(0, min(255, x + random.randint(-self.range, self.range))))
 
   # Constructs a string representation of this Augmentation.
   def __repr__(self):
-    return f"JpegAugmentation(quality_range={self.quality_range}, probability={self.probability})"
+    return f"SubtleNoiseAugmentation(range={self.range}, probability={self.probability})"
 
   # Applies the Augmentation to input data.
   def __call__(self, data, force=False):
     if (force or self.should_run()):
       image = data['post'][-1].result
-      encode_param = [int(cv2.IMWRITE_JPEG_QUALITY), random.randint(self.quality_range[0], self.quality_range[1])]
-      result, encimg = cv2.imencode('.jpg', image, encode_param)
-      image = cv2.imdecode(encimg, 1)
-      data['post'].append(AugmentationResult(self, image))
+      data['post'].append(AugmentationResult(self, self.add_subtle_noise(image)))
+
